+++ src/helpers/buildWalletSections.js
@@ -132,7 +132,7 @@ const withUniswapSection = (
   return {
     data: uniswap,
     header: {
-      title: 'Pools',
+      title: lang.t('account.tab_investments'),
       totalItems: uniswap.length,
       totalValue: convertAmountToNativeDisplay(uniswapTotal, nativeCurrency),
     },
@@ -225,7 +225,6 @@ const coinEditContextMenu = (
   sortedAssets,
   balanceSectionData,
   isCoinListEdited,
-  currentAction,
   isLoadingAssets,
   sortedAssetsCount,
   totalValue,
@@ -239,7 +238,7 @@ const coinEditContextMenu = (
         ? {
             cancelButtonIndex: 0,
             dynamicOptions: () => {
-              return ['Cancel', 'Edit'];
+              return [lang.t('button.cancel'), lang.t('button.edit')];
             },
             onPressActionSheet: async index => {
               if (index === 1) {