+++ src/parsers/uniqueTokens.js
@@ -88,7 +88,7 @@ export const parseAccountUniqueTokens = data => {
         urlSuffixForAsset: `${get(asset_contract, 'address')}/${token_id}`,
       })
     )
-    .filter(token => !!token.familyName && token.familyName !== 'POAP');
+    .filter(token => !!token.familyName);
 };
 
 export const getFamilies = uniqueTokens =>