+++ src/redux/settings.ts
@@ -1,10 +1,10 @@
 import analytics from '@segment/analytics-react-native';
 import { Dispatch } from 'redux';
 import { ThunkDispatch } from 'redux-thunk';
-import { updateLanguage } from '../languages';
-import { saveString } from '../model/keychain';
+import { updateLanguageLocale } from '../languages';
 import { NativeCurrencyKeys } from '@rainbow-me/entities';
 import {
+  getLanguage,
   getNativeCurrency,
   getNetwork,
   saveLanguage,
@@ -106,6 +106,21 @@ export const settingsLoadNetwork = () => async (
   }
 };
 
+export const settingsLoadLanguage = () => async (
+  dispatch: Dispatch<SettingsStateUpdateLanguageSuccessAction>
+) => {
+  try {
+    const language = await getLanguage();
+    updateLanguageLocale(language);
+    dispatch({
+      payload: language,
+      type: SETTINGS_UPDATE_LANGUAGE_SUCCESS,
+    });
+  } catch (error) {
+    logger.log('Error loading language settings', error);
+  }
+};
+
 export const settingsUpdateAccountAddress = (accountAddress: string) => async (
   dispatch: Dispatch<SettingsStateUpdateSettingsAddressAction>
 ) => {
@@ -134,7 +149,7 @@ export const settingsUpdateNetwork = (network: Network) => async (
 export const settingsChangeLanguage = (language: string) => async (
   dispatch: Dispatch<SettingsStateUpdateLanguageSuccessAction>
 ) => {
-  updateLanguage(language);
+  updateLanguageLocale(language);
   try {
     dispatch({
       payload: language,
@@ -163,10 +178,6 @@ export const settingsChangeNativeCurrency = (nativeCurrency: string) => async (
     });
     dispatch(explorerInit());
     saveNativeCurrency(nativeCurrency);
-    saveString('nativeCurrency', nativeCurrency, {
-      accessGroup: 'group.rainbow.me',
-      service: 'rainbow.me.currency',
-    });
     analytics.identify(null, { currency: nativeCurrency });
   } catch (error) {
     logger.log('Error changing native currency', error);