+++ src/screens/BackupSheet.js
@@ -144,25 +144,30 @@ export default function BackupSheet() {
       case WalletBackupStepTypes.existing_user:
         return (
           <BackupSheetSection
-            descriptionText="You have wallets that have not been backed up yet. Back them up in case you lose this device."
+            descriptionText={lang.t('modal.back_up.existing.description')}
             onPrimaryAction={onBackupNow}
             onSecondaryAction={goBack}
-            primaryLabel="Back up now"
-            secondaryLabel="Maybe later"
-            titleText="Would you like to back up?"
+            primaryLabel={lang.t('modal.back_up.existing.button.now')}
+            secondaryLabel={lang.t('modal.back_up.existing.button.later')}
+            titleText={lang.t('modal.back_up.existing.title')}
             type="Existing User"
           />
         );
       case WalletBackupStepTypes.imported:
         return (
+          //TODO: ADD CloudPlatform to back_up.description
           <BackupSheetSection
-            descriptionText={`Don't lose your wallet! Save an encrypted copy to ${cloudPlatform}.`}
+            descriptionText={lang.t('modal.back_up.imported.description', {
+              cloudPlatformName: cloudPlatform,
+            })}
             onPrimaryAction={onIcloudBackup}
             onSecondaryAction={goBack}
-            primaryLabel={`􀙶 Back up to ${cloudPlatform}`}
+            primaryLabel={`􀙶 ${lang.t('modal.back_up.imported.button.back_up', {
+              cloudPlatformName: cloudPlatform,
+            })}`}
             secondaryButtonTestId="backup-sheet-imported-cancel-button"
-            secondaryLabel="No thanks"
-            titleText="Would you like to back up?"
+            secondaryLabel={lang.t('modal.back_up.imported.button.no_thanks')}
+            titleText={lang.t('modal.back_up.imported.title')}
             type="Imported Wallet"
           />
         );
@@ -177,12 +182,18 @@ export default function BackupSheet() {
       default:
         return (
           <BackupSheetSection
-            descriptionText={`Don't lose your wallet! Save an encrypted copy to ${cloudPlatform}.`}
+            descriptionText={lang.t('modal.back_up.default.description', {
+              cloudPlatformName: cloudPlatform,
+            })}
             onPrimaryAction={onIcloudBackup}
             onSecondaryAction={onManualBackup}
-            primaryLabel={`􀙶 Back up to ${cloudPlatform}`}
-            secondaryLabel="🤓 Back up manually"
-            titleText="Back up your wallet"
+            primaryLabel={`􀙶 ${lang.t('modal.back_up.default.button.cloud', {
+              cloudPlatformName: cloudPlatform,
+            })}`}
+            secondaryLabel={`🤓 ${lang.t(
+              'modal.back_up.default.button.manual'
+            )}`}
+            titleText={lang.t('modal.back_up.default.title')}
             type="Default"
           />
         );